docs: clarify build step in contributing guide
